+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 8.6.0
+  * Upgrade to Font Awesome Icons 5.11.2
+  * Add `values` array that lists all icons.
+  * Add `final` keyword where possible.
+  * Set minimal Dart SDK version to v2.2.2
+
 ## 8.5.0
   * Upgrade to Font Awesome Icons 5.9